Use regex in order to filter out bad package names.
Fix the issue of improper loading of folders (e.g. '.DS_Store')
by filtering out all the files/directories that doesn't match with
a package Python name.

Why? Because if that's not the case, the component cannot be used in a 'import' statement.
In the future, there might be the need to have a 'ignore' file with all the known patterns whose matching string shouldn't be considered correct.