fix loading of modules for shared classes finding.
The old approach was implemented in the wrong way.
It tried to load any kind of file that wasn't one of
behvaiours.py, handlers.py, tasks.py and skill.yaml.

Now, globbing with the pattern "*.py" is used, and the result
filtered if the string match the filenames mentioned before.