customize click.testing.CLIRunner and update tests accordingly.
Copy the `click.testing` module into the tests.common of the project.
Update a part of the code in that module to overcome an issue
during testing click applications. Intuitively, it seems that
the click testing framework tries to read the (isolated) outstream
(hence not the true sys.stdout) even if it has been closed (maybe due to sys.exit?).
So just add a check whether the outstream has been closed; if not,
we can read, otherwise just skip the read.