update Agent.stop method with new outbox.put_message semantics.
After the change from MailBox to Multiplexer, OutBox.put
and OutBox.put_message are blocking methods if the
multiplexer is not running.
However, that casues issues when a skill in a teardown method
of some Behaviour tries to use the connection to send the last
messages, as it happened in #426.

To fix this problem, the Agent.start and Agent.stop should
initialize and tear down resources in a mirrored way.
That is:

Agent.start():
    1) set connections up
    2) set skills up
    3) liveness to True

Agent.stop():
    3) liveness to False
    2) tear skills down
    1) tear connections down

That wasn't the case before this commit. Indeed, the Agent.stop()
method was:

Agent.stop():
    3) liveness to False
    1) tear connections down
    2) tear skills down

But if the skills for some reason need the connection up
in order to release the resources correctly, this approach might fail.