remove 'LedgerAPIConfig' class. Update ledger_apis configuration schema.
LedgerAPIConfig appeared to be of little utility.
The 'ledger_apis:' field in 'aea-config.yaml' configuration file
contains key-value pair where the key is the ledger api identifier
and the value is an object. Previously, it contained the key
'args' whose value was the actual dictionary of arguments.

E.g. before, it was like the following:

ledger_apis:
   fetchai:
     args:
       addr: example.fetch-ai.com
       port: 8080
   ethereum:
	 args:
       addr: example.ethereum.com

Now:

ledger_apis:
   fetchai:
     addr: example.fetch-ai.com
     port: 8080
   ethereum:
     addr: example.ethereum.com