update echo skill.
After the recent changes on the task system,
the echo skill does not load the Echo Task automatically.
Hence, since the default 'timeout' for an AEA agent is 0.0,
the act method of the EchoBehaviour gets called many times.

To solve this problem, make 'EchoBehaviour' inherit from 'TickerBehaviour',
with the 'tick_interval' default value set in the associated 'skill.yaml' file.

TODO: the docs on the Echo demo must be updated.