make http connection use the already-running asyncio loop
the assumption underlying the AEA connections is that the
attribute 'self.loop' is already running, because it is started by
the multiplexer. (notice, 'self.loop' of the connection is the same loop
passed to the http channel).
In the tests, if the multiplexer is not used, the loop must be run in
a thread (or, even better, using the asyncio loop being run by the
pytest-asyncio plugin). That's the reason why the code with
'run_coroutine_threadsafe' didn't work.
Moreover, the approach of 'run_until_complete' won't work with the
multiplexer, because at the time the method is called the loop
is already running:

    RuntimeError: This event loop is already running

Change (only one) test to show the above in code.