add support for more than one specifier.
the regular expression of 'version_specifier' in 'definition.json' jsonschema
has been extended to support many specifiers.

In particular, the regex was like the following:

    ^(specifier)?$

It used to match a single specifier, that could also be empty.

Now, we have the following:

    ^(specifier)(,\s*specifier)*

That is, we match at least one specifier, plus an unbounded
sequence of specifiers, comma-separated with optional spaces in between.

The same pattern supports both the 'dependencies' field
and the 'aea_version' field in the configuation files.