add recommended order for fields in yaml files.
Now the configuration files, when created, have the fields
in a "recommended" order. They have 'name', 'author', 'version',
'description', 'license' and other basic details  at the beginning.
Other fields have different order, depending on the type of
configuration.

More importantly, the user can decide the order, and the framework ensures
that it will be kept.