add AEAException class.
This class is a custom exception for the AEA framework.
It is supposed to be used whenever we need to describe
exceptions at framework level.

For instance, trying to add a package in the AEABuilder class
(`AEABuilder.add_component(type, path)`) might fail for many reasons:

1. the path to the package does not exists
2. the directory exists, but there isn't the expected configuration file.
3. the package already exists in the AEABuilder
4. there are some missing package dependencies
...

Desipite we can't foresee all the exceptions, for some of
them we can. For instance, case (3) and (4) are exception
at framework level, so we can raise the right exception to
the client of the class.

Needless to say, AEABuilder.add_component' is just one example.
The same idea can be applied to other framework classes/functions.