allow for deregistration of skill in skill component setup().
Changes:

1) instantiate skill object before skill component loading
2) in skill component registries, setup() now dispatches the call
   to only the skill components whose skill is active.

1)

The main motivation for this change is due to
https://github.com/fetchai/agents-aea/issues/1095
In the tac_controller_contract skill, the Parameter model
checks whether the skill should be deactivated or not
in its __init__ method. But if the condition is met,
it will try to set the attribute context.is_active,
which in turn requires the skill object being available
to retrieve the skill id for a debug log message.

Hence, before this commit, trying to deactivate a skill from the setup()
method of some skill component would have failed.

The reason is that the current implementation of the loading of a skill
follows this order:

Skill.from_config(configuration, skill_context):
    # notice: skill_context._skill is None
    load handlers/behaviours/models
    # instantiate skill object
    skill = Skill(configuration, skill_context, handlers, behaviours, models)
    # set the skill instance to the skill context
    skill_context._skill = skill
    return skill

However, if a skill component needed to deactivate the skill, i.e.

   context.is_active = False

this would have caused an error because the setter of the attribute
'is_active' needs the skill_id (and so the skill object)
in order to log a debug message, which was not available
since the skill object was instantiated at the end of the loading of each component.

In order to fix that, move the instantiation before the loading of each skill,
and add the skill components to the skill object at the end, i.e.

Skill.from_config(configuration, skill_context):
    # this will instantiate empty indexes of skill components
    skill = Skill(configuration, skill_context)
    skill._skill = skill
    load handlers/behaviours/models
    skill.handlers.update(handlers)
    skill.behaviours.update(behaviours)
    skill.models.update(models)
    return skill

2)

Due to the previous change, it could happen that some
skills are deactivated before the setup() methods
of the skill components are called.

This change will make the component registry
to first check whether a skill associated to a component
is active before calling the setup() method of the component.

More work needs to be done to hanle re-activation and
re-calling of setup(), as well as teardown of deactivated skill.