avoid using google.protobuf.Any
it results in too much overhead with no relevant benefits;
Instead, we decided to use the type bytes for DialogueMessage.content field.

The result is that the dialogue fields act as message header,
whereas the actual performative content is serialized as bytes, in a
format specified by the protocol serialization layer.

This will give more flexibility for protocol implementations.