fix a bug on 'aea push --local' with 'latest' as version
The problem was the comparison between the provided public id
and the existing local package public id
the 'aea push' CLI command, when the '--local' flag was provided
and the public id provided as argument contained ':latest'
(or equivalently '') as version number.
Now if latest is provided, only the public id prefix is compared,
o/w also the version is compared.