fix: add consistency check in 'aea install'
before this change, the problem was that if there were
two PyPI dependencies with the same package,
one of them replaced the other in an unpredictable way.

Now instead we check upfront whether there are two dependencies
with the same package name, and we also check whether they are
conflicting or unsatisfiable.